[SH-20735] Box Button > Small 사이즈 상세 스팩 변경 (2.1.35)
